Create Main
This code defines a Invoice class that can be used to generate invoices. The Invoice class has a constructor that takes in the invoice number, customer name, and amount due, and uses these values to initialize the instance variables. It also has a generateInvoice() method that generates an invoice string using the instance variables and returns it as the output.

To use this code, you can create an Invoice object and use the generateInvoice() method to generate an invoice string. You can then print the invoice string or save it to a file, as needed.